added individual index.js files for folders and subcomponents, moved edit-ad files into their own folder
