+++ packages/p2p/src/pages/my-ads/ad-status/__tests__/ad-status.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { screen, render } from '@testing-library/react';
-import AdStatus from '../ad-status.jsx';
+import AdStatus from '../ad-status';
 
 describe('<AdStatus />', () => {
     it('Component should show proper status if is_active prop is true', () => {
