+++ packages/p2p/src/pages/my-ads/ad-status/index.js
@@ -1,0 +1,3 @@
+import AdStatus from './ad-status';
+
+export default AdStatus;
