+++ packages/p2p/src/pages/my-ads/ad-type/index.js
@@ -1,0 +1,3 @@
+import AdType from './ad-type';
+
+export default AdType;
