+++ packages/p2p/src/pages/my-ads/buy-ad-payment-methods-list/index.js
@@ -1,0 +1,3 @@
+import BuyAdPaymentMethodsList from './buy-ad-payment-methods-list';
+
+export default BuyAdPaymentMethodsList;
