+++ packages/p2p/src/pages/my-ads/create-ad/create-ad-form-payment-methods/index.js
@@ -1,0 +1,3 @@
+import CreateAdFormPaymentMethods from './create-ad-form-payment-methods';
+
+export default CreateAdFormPaymentMethods;
