+++ packages/p2p/src/pages/my-ads/create-ad/create-ad-form/index.js
@@ -1,0 +1,3 @@
+import CreateAdForm from './create-ad-form.jsx';
+
+export default CreateAdForm;
