+++ packages/p2p/src/pages/my-ads/create-ad/create-ad-summary/index.js
@@ -1,0 +1,3 @@
+import CreateAdSummary from './create-ad-summary';
+
+export default CreateAdSummary;
