+++ packages/p2p/src/pages/my-ads/create-ad/create-ad.jsx
@@ -4,7 +4,7 @@ import { observer } from 'mobx-react-lite';
 import { localize } from 'Components/i18next';
 import PageReturn from 'Components/page-return/page-return.jsx';
 import { useStores } from 'Stores';
-import CreateAdForm from './create-ad-form/create-ad-form.jsx';
+import CreateAdForm from './create-ad-form';
 
 const CreateAd = () => {
     const { my_ads_store } = useStores();
