+++ packages/p2p/src/pages/my-ads/create-ad/index.js
@@ -1,0 +1,3 @@
+import CreateAd from './create-ad.jsx';
+
+export default CreateAd;
