+++ packages/p2p/src/pages/my-ads/edit-ad/edit-ad-form-payment-methods/index.js
@@ -1,0 +1,3 @@
+import EditAdFormPaymentMethods from './edit-ad-form-payment-methods';
+
+export default EditAdFormPaymentMethods;
