+++ packages/p2p/src/pages/my-ads/edit-ad/edit-ad-form/edit-ad-form.jsx
@@ -14,9 +14,9 @@ import { ad_type } from 'Constants/floating-rate';
 import FloatingRate from 'Components/floating-rate';
 import EditAdCancelModal from 'Components/my-ads/edit-ad-cancel-modal.jsx';
 import { generateErrorDialogTitle, generateErrorDialogBody } from 'Utils/adverts';
-import EditAdFormPaymentMethods from './edit-ad-form-payment-methods/edit-ad-form-payment-methods.jsx';
+import EditAdFormPaymentMethods from '../edit-ad-form-payment-methods';
 import CreateAdAddPaymentMethodModal from 'Components/my-ads/create-ad-add-payment-method-modal.jsx';
-import EditAdSummary from './edit-ad-summary.jsx';
+import EditAdSummary from '../edit-ad-summary';
 
 const EditAdFormWrapper = ({ children }) => {
     if (isMobile()) {
