+++ packages/p2p/src/pages/my-ads/edit-ad/edit-ad-form/index.js
@@ -1,0 +1,3 @@
+import EditAdForm from './edit-ad-form';
+
+export default EditAdForm;
