+++ packages/p2p/src/pages/my-ads/edit-ad/edit-ad-summary/index.js
@@ -1,0 +1,3 @@
+import EditAdSummary from './edit-ad-summary';
+
+export default EditAdSummary;
