+++ packages/p2p/src/pages/my-ads/edit-ad/index.js
@@ -1,0 +1,3 @@
+import EditAd from './edit-ad.jsx';
+
+export default EditAd;
